Create a file for all constants to be tested
